+++ include/cis_scm/rgbd_driver.hpp
@@ -26,16 +26,14 @@
 #include <sensor_msgs/msg/point_cloud2.hpp>
 #include <camera_info_manager/camera_info_manager.hpp>
 
-#ifndef INTERNAL_DRIVER
-#include "cis_scm/Device.h"
-#else
-#include "cis_scm/InternalDevice.hpp"
-#endif
 #include "cis_scm/tof_driver.hpp"
 
 namespace cis_scm
 {
 
+/**
+ * @brief Combined 3D data structure with XYZ coordinates and RGB values. Also placeholder for ir data
+ */
 struct XYZRGBData
 {
     XYZData xyz;
@@ -43,6 +41,9 @@ struct XYZRGBData
     std::vector<uint8_t> ir;
 };
 
+/**
+ * @brief Single RGB-D pixel with spatial and color data.
+ */
 struct XYZRGBPixel
 {
     float x;
@@ -52,11 +53,25 @@ struct XYZRGBPixel
     uint8_t a;
 };
 
-class RGBDNode : public ToFCVNode
+/**
+ * @brief ROS2 node for SCM-RGBD1 device, combining Depth and RGB nodes
+ *
+ * Extends ToFNode to handle synchronized RGB and depth frames,
+ * producing aligned RGB-D images and point clouds.
+ */
+class RGBDNode : public ToFNode
 {
   public:
+    /**
+   * @brief Construct a new RGB-D node.
+   * @param node_name ROS node name.
+   * @param node_options Node options.
+   */
     RGBDNode(const std::string node_name, const rclcpp::NodeOptions & node_options);
 
+    /**
+   * @brief Start the RGB-D capture and publication loop.
+   */
     void start() override;
 
   protected:
@@ -65,10 +80,6 @@ class RGBDNode : public ToFCVNode
     sensor_msgs::msg::Image imgRGBMsg_;
     bool isPCLNoColor_;
 
-#ifdef INTERNAL_DRIVER
-    std::unique_ptr<internal::RGBDInternalDevice> cap_;
-#endif
-
     std::shared_ptr<rclcpp::Publisher<sensor_msgs::msg::Image>> rgbImgPub_;
     std::shared_ptr<rclcpp::Publisher<sensor_msgs::msg::CameraInfo>> infoRGBPub_;
     std::shared_ptr<camera_info_manager::CameraInfoManager> cinfo_rgb_;
@@ -76,6 +87,10 @@ class RGBDNode : public ToFCVNode
     std::string cameraColorFrame_ = "camera_color_frame";
     std::string topicRGBPrefix_ = "aligned_color_to_depth/";
 
+    /**
+   * @brief Initialize the RGB-D capture device.
+   * @return 0 on success, negative on failure.
+   */
     int initCap() override;
     void importRGBDParameters();