+++ src/rgb_driver.cpp
@@ -19,13 +19,7 @@
 #include <sensor_msgs/image_encodings.hpp>
 #include <sensor_msgs/point_cloud2_iterator.hpp>
 
-#ifndef INTERNAL_DRIVER
 #include "cis_scm/ExternalDevice.hpp"
-#else
-#include <image_transport/image_transport.hpp>
-
-#include "cis_scm/InternalDevice.hpp"
-#endif
 
 using namespace std::literals::chrono_literals;
 namespace cis_scm
@@ -38,9 +32,7 @@ RGBNode::RGBNode(const std::string node_name, const rclcpp::NodeOptions & node_o
 
     cinfo_ = std::make_shared<camera_info_manager::CameraInfoManager>(this);
 
-#ifndef INTERNAL_DRIVER
     imgPub_ = create_publisher<sensor_msgs::msg::Image>(topicRGBPrefix_ + "image", 10);
-#endif
     importParams();
 }
 
@@ -87,27 +79,14 @@ void RGBNode::importParams()
     }
 }
 
-#ifdef INTERNAL_DRIVER
-void RGBNode::initImageTransport()
-{
-    it_ = std::make_shared<image_transport::ImageTransport>(shared_from_this());
-    imgPub_ = it_->advertise(topicPrefix_ + "/image_raw", 10);
-}
-#endif
-
 int RGBNode::initCap()
 {
-#ifdef INTERNAL_DRIVER
-    cap_ = std::make_unique<internal::RGBInternalDevice>(new cis::CameraAR0234());
-    if (!cap_->connect())
-#else
     cap_ = std::make_unique<ExternalDevice>();
-    if (!cap_->connect(width_, height_))
-#endif
-    {
+    if (!cap_->connect(width_, height_)) {
         RCLCPP_INFO(get_logger(), "Camera connected");
     } else {
         RCLCPP_ERROR(get_logger(), "Camera connection failed");
+        return -1;
     }
     return 0;
 }
@@ -128,7 +107,7 @@ void RGBNode::start()
         rclcpp::shutdown();
     }
 
-    if (cap_->isConnected()) {
+    if (cap_->isStreamOn()) {
         DevInfo devInfo = cap_->getInfo();
         dispInfo(devInfo);
         width_ = devInfo.width;
@@ -158,17 +137,12 @@ void RGBNode::pubImage(uint8_t * data)
     imgMsg_.step = width_ * 3;
     imgMsg_.is_bigendian = true;
     imgMsg_.data.assign(data, data + imgMsg_.step * imgMsg_.height);
-#ifdef INTERNAL_DRIVER
-    imgPub_.publish(std::move(imgMsg_));
-#else
     imgPub_->publish(std::move(imgMsg_));
-#endif
 }
 
 void RGBNode::imgCallback()
 {
-    auto st = this->now();
-    if (!cap_->isConnected()) {
+    if (!cap_->isStreamOn()) {
         RCLCPP_ERROR(get_logger(), "Camera connection lost or unavailable");
         rclcpp::sleep_for(std::chrono::seconds(3));
         RCLCPP_INFO(get_logger(), "New camera connection attempt");
@@ -187,7 +161,6 @@ void RGBNode::imgCallback()
             pubImage(imgData.data);
         }
     }
-    auto sp = this->now();
 }
 
 }  // namespace cis_scm