+++ src/tof_driver.cpp
@@ -26,22 +26,14 @@
 #include <sensor_msgs/point_cloud2_iterator.hpp>
 #include <camera_info_manager/camera_info_manager.hpp>
 
-#ifndef INTERNAL_DRIVER
 #include "cis_scm/Device.h"
-#else
-#include <rmw/types.h>
-
-#include <point_cloud_transport/point_cloud_transport.hpp>
-
-#include "cis_scm/InternalDevice.hpp"
-#endif
 
 using namespace std::chrono_literals;
 
 namespace cis_scm
 {
 
-ToFCVNode::ToFCVNode(const std::string node_name, const rclcpp::NodeOptions & node_options)
+ToFNode::ToFNode(const std::string node_name, const rclcpp::NodeOptions & node_options)
 : Node(node_name, node_options)
 
 {
@@ -52,53 +44,27 @@ ToFCVNode::ToFCVNode(const std::string node_name, const rclcpp::NodeOptions & no
     infoPub_ =
         create_publisher<sensor_msgs::msg::CameraInfo>(topicDepthPrefix_ + "camera_info", 10);
 
-#ifndef INTERNAL_DRIVER
     depthPCLPub_ =
         create_publisher<sensor_msgs::msg::PointCloud2>(topicDepthPrefix_ + "points", 10);
-#endif
 
     cinfo_ = std::make_shared<camera_info_manager::CameraInfoManager>(this, "scm-tof1");
 
     importParams();
 }
 
-#ifdef INTERNAL_DRIVER
-void ToFCVNode::initPointCloudTransport()
+int ToFNode::initCap()
 {
-    // Humble ver. of point_cloud_transport does not use rclcpp QoS yet
-    rmw_qos_profile_t pc_qos = rmw_qos_profile_sensor_data;
-    pc_qos.depth = 30;
-    // Reliable because too many fragmented dataframe
-    pc_qos.reliability = RMW_QOS_POLICY_RELIABILITY_RELIABLE;
-    pc_qos.durability = RMW_QOS_POLICY_DURABILITY_VOLATILE;
-    pc_qos.history = RMW_QOS_POLICY_HISTORY_KEEP_LAST;
-
-    depthPCLPub_ = point_cloud_transport::create_publisher(
-        shared_from_this(), topicPrefix_ + "/pcl_depth", pc_qos);
-}
-#endif
-
-int ToFCVNode::initCap()
-{
-#ifndef INTERNAL_DRIVER
-    RCLCPP_INFO(get_logger(), "EXTERNAL DRIVERRRR");
     cap_ = std::make_unique<ExternalDevice>();
-    if (!cap_->connect(480, 640))
-#else
-    RCLCPP_INFO(get_logger(), "INTERNAL DRIVERRRR");
-    cap_ = std::make_unique<internal::ToFInternalDevice>(
-        this->get_parameter("tof_params_path").as_string());
-    if (!cap_->connect())
-#endif
-    {
+    if (!cap_->connect(480, 640)) {
         RCLCPP_INFO(get_logger(), "Camera connected");
     } else {
         RCLCPP_ERROR(get_logger(), "Camera connection failed");
+        return -1;
     }
     return 0;
 }
 
-void ToFCVNode::dispInfo(DevInfo devInfo) const
+void ToFNode::dispInfo(DevInfo devInfo) const
 {
     RCLCPP_INFO(get_logger(), "Device name: %s", devInfo.devName.c_str());
     RCLCPP_INFO(get_logger(), "Driver version: %s", devInfo.driverVers.c_str());
@@ -107,7 +73,7 @@ void ToFCVNode::dispInfo(DevInfo devInfo) const
     RCLCPP_INFO(get_logger(), "Height: %d", devInfo.height);
 }
 
-void ToFCVNode::importParams()
+void ToFNode::importParams()
 {
     // Cam params
     if (!this->has_parameter("tof_camera_params")) {
@@ -134,14 +100,14 @@ void ToFCVNode::importParams()
     }
 }
 
-void ToFCVNode::start()
+void ToFNode::start()
 {
     if (initCap()) {
         RCLCPP_ERROR(get_logger(), "Camera initialization failed");
         rclcpp::shutdown();
     }
 
-    if (cap_->isConnected()) {
+    if (cap_->isStreamOn()) {
         DevInfo devInfo = cap_->getInfo();
         dispInfo(devInfo);
         width_ = 640;
@@ -157,14 +123,14 @@ void ToFCVNode::start()
 
         RCLCPP_INFO(get_logger(), "Start capturing");
         timer_ = this->create_wall_timer(
-            66.666ms, std::bind(&ToFCVNode::depthCallback, this), crit_cb_grp_);
+            66.666ms, std::bind(&ToFNode::depthCallback, this), crit_cb_grp_);
     } else {
         RCLCPP_ERROR(get_logger(), "Camera connection failed");
         rclcpp::shutdown();
     }
 }
 
-XYZData ToFCVNode::splitXYZ(float * data)
+XYZData ToFNode::splitXYZ(float * data)
 {
     int size = width_ * height_;
     XYZData output;
@@ -179,17 +145,10 @@ XYZData ToFCVNode::splitXYZ(float * data)
     return output;
 }
 
-void ToFCVNode::pubDepthImage(float * data)
+void ToFNode::pubDepthImage(float * data)
 {
-    greyDepth_ = cv::Mat(height_, width_, CV_32FC1, data);
-    greyDepth_.convertTo(greyDepth_, CV_8UC1, 255. / MAX_DEPTH);
-
-    // Grey -> Hue
-    hueDepth_ = cv::Mat(height_, width_, CV_8UC3);
-    cv::applyColorMap(greyDepth_, hueDepth_, cv::COLORMAP_JET);
-    if (!hueDepth_.isContinuous()) {
-        printf("NOT CONTINOUS");
-    }
+    depthMap_ = cv::Mat(height_, width_, CV_32FC1, data);
+    depthMap_.convertTo(depthMap_, CV_8UC1, 255. / MAX_DEPTH);
 
     // Msg header
     auto header = std_msgs::msg::Header();
@@ -201,14 +160,14 @@ void ToFCVNode::pubDepthImage(float * data)
     imgMsg_.header = header;
     imgMsg_.width = width_;
     imgMsg_.height = height_;
-    imgMsg_.encoding = sensor_msgs::image_encodings::TYPE_8UC3;
-    imgMsg_.step = width_ * sizeof(uchar) * 3;
+    imgMsg_.encoding = sensor_msgs::image_encodings::TYPE_8UC1;
+    imgMsg_.step = width_ * sizeof(uchar);
     imgMsg_.is_bigendian = true;
-    imgMsg_.data.assign(hueDepth_.data, hueDepth_.data + hueDepth_.rows * hueDepth_.cols * 3);
+    imgMsg_.data.assign(depthMap_.data, depthMap_.data + depthMap_.rows * depthMap_.cols);
     depthImgPub_->publish(std::move(imgMsg_));
 }
 
-void ToFCVNode::pubDepthPtc(XYZData & data)
+void ToFNode::pubDepthPtc(XYZData & data)
 {
     // Header
     auto header = std_msgs::msg::Header();
@@ -234,20 +193,16 @@ void ToFCVNode::pubDepthPtc(XYZData & data)
             *iter_z = z;
         }
     }
-#ifdef INTERNAL_DRIVER
-    depthPCLPub_.publish(ptcMsg_);
-#else
     depthPCLPub_->publish(ptcMsg_);
-#endif
 }
 
-void ToFCVNode::depthCallback()
+void ToFNode::depthCallback()
 {
     // Init 2D Mats
 
     std::vector<float> frameData = std::vector<float>(width_ * height_ * 3);
 
-    if (!cap_->isConnected()) {
+    if (!cap_->isStreamOn()) {
         RCLCPP_ERROR(get_logger(), "Camera connection lost or unavailable");
         rclcpp::sleep_for(std::chrono::seconds(3));
         RCLCPP_INFO(get_logger(), "New camera connection attempt");